move GraphicsIO.set_x() to Memory
It makes sense for the set_byte/hw/word methods of each IO struct to be able to read from Memory using Memory.get_byte/hw/word, but it's impossible to call self.graphics.set_byte(addr, val, &self) from inside Memory because of borrow errors (even if we could it would be a weird way to do things). So move the method to Memory instead and call it update_graphics_byte/hw/word. Not sure if this is the best way to do things but i'll stick with it for now.